ICP Final Run

+++ mattest.m
@@ -14,7 +14,7 @@ function mattest( jobID, taskID )
     % 5: ICP1
     % 6: Chamfer
 
-    algo = 5;
+    algo = 4;
 
 
     % Set up paths
@@ -30,11 +30,19 @@ function mattest( jobID, taskID )
                  [wd '/util']);
 
     % Output Path
-    OutPath = ['../' num2str(jobID) '/'];
+    OutPath = ['../' num2str(jobID) '/' num2str(taskID, '%3u') '-'];
     mkdir(OutPath);
 
     % Keep a copy of the settings with the results
-    copyfile('SLAM.m', OutPath)
+    pause(taskID);
+    if ~exist([OutPath 'SLAM.m'], 'file')
+      copyfile('SLAM.m'   , OutPath)
+      copyfile('mattest.m', OutPath)
+    end
+
+    %Default dataset ROI
+    start         = 1;     % Scan Index
+    stop          = 100000; % Scan Index
 
     % Load sensor dataset
     %load('testData/testWorldPSM.mat')
@@ -46,31 +54,28 @@ function mattest( jobID, taskID )
         case 1
             DatasetName = 'hall_and_room_w_vn';
         case 2
-            DatasetName = 'eerc_dillman_dow';
+            DatasetName = 'eerc_dillman_dow';  
+            %stop        = 20000;
         case 3
-            DatasetName = 'campus1';  
+            DatasetName = 'campus1'; 
+            %stop        = 25000;
         case 4
             DatasetName = 'eerc_dow_dill_inout';  % Bad timestamp early in dataset {20}
+            start       = 20;   
         case 5
-            DatasetName = '2000-01-31-19-01-35';  % EERC817 Multiple passes
-        case 6
             DatasetName = '2000-01-31-19-05-32';  % EERC8F
-        case 7
-            DatasetName = '2000-01-31-19-10-14';  % EERC1F
-        case 8
-            DatasetName = '2000-01-31-19-13-51';  % EERC1F small loop 
-        case 9
-            DatasetName = '2000-01-31-19-15-35';  % EERC1F large loop
-        case 10
+        case 6
             DatasetName = '2000-01-31-19-21-26';  % EERC_DOW_DIL inout  [1-32000],
-        case 11
+        case 7
             DatasetName = '2000-01-31-19-50-23';  % Campus {1.5, 16k, 24k, 43k, [], 76k, 90k} 
-        case 12
+        case 8
             DatasetName = '2000-01-31-20-30-59';  % EERC8f elevator to ieee, 817
-        case 13
-            DatasetName = 'EERC8f handheld';      % EERC 8f IEEE, Lap, stairwells. (good)
-        case 14
-            DatasetName = 'EERC_DOW_DIL inout3';  % Handheld {1, 3k, 11k, 24k }
+        case 9
+            DatasetName = 'EERC8f handheld';      % EERC 8f IEEE, Lab, stairwells. (good)
+            start       = 500;   
+        case 10
+            DatasetName = 'EERC_DOW_DIL inout3';  % Handheld {EERC8F 1, EERC 1F 3k, 11k, 24k }
+            start       = 3000;   
     end
 
 
@@ -86,6 +91,12 @@ function mattest( jobID, taskID )
 
     SLAM
 
-    save([ OutPath DatasetName '.mat'], '-v7.3');
+    % Remove debug plot when finished
+    if exist([ OutPath DatasetName '-dbg.pdf' ], 'file')
+      delete[ OutPath DatasetName '-dbg.pdf' ]
+    end
+    
+    % Save result
+    %save([ OutPath DatasetName '.mat'], '-v7.3');
 end
